% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{byte2hex}
\alias{byte2hex}
\title{Get a two-character hex string representation of a number.}
\usage{
byte2hex(byte)
}
\arguments{
\item{byte}{Integer value in range [0,255]}
}
\value{
String representation of byte.
}
\description{
Get a two-character hex string representation of a number.
}
\keyword{internal}
