% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp}
\alias{clamp}
\title{Constrain a numeric value within a range.}
\usage{
clamp(value, min.value, max.value)
}
\arguments{
\item{value}{Value to constrain.}

\item{min.value}{Minimum, inclusive.}

\item{max.value}{Maximum, inclusive.}
}
\value{
The closest number in the range [min.value,max.value] to value.
}
\description{
Constrain a numeric value within a range.
}
