% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{column.to.vector}
\alias{column.to.vector}
\title{Get a vector from a data frame column.}
\usage{
column.to.vector(data, column)
}
\arguments{
\item{data}{Data frame.}

\item{column}{Column to extract - can be used with \{\{\}\}.}
}
\value{
Vector of column contents.
}
\description{
Get a vector from a data frame column.
}
