% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_reps.R
\name{counts.to.reps}
\alias{counts.to.reps}
\title{Transform pass and fail counts to repeated rows.}
\usage{
counts.to.reps(data, pass.col, fail.col, result.col)
}
\arguments{
\item{data}{\code{data.frame} with columns \code{pass.col} and \code{fail.col}.}

\item{pass.col}{\code{integer} count of individuals passing some binary test}

\item{fail.col}{\code{integer} count of individuals failing some binary test}

\item{result.col}{name to use for result column}
}
\value{
A \code{data.frame} with counts represented as repeated rows with a
\code{logical} flag representing a pass or fail.
}
\description{
Given a data frame with a column representing the count \emph{p} of individuals
that passed some binary test, and another column representing the count \emph{q}
of individuals that failed the test, return a new data frame where each row
in the original is repeated \emph{p} times with a \code{TRUE} value in a new
\code{result.col} column and \emph{q} times with a \code{FALSE} value in that column.
}
\examples{
d <- tibble::tribble(~treatment, ~day, ~metamorphosed, ~did.not.metamorphose,
   "A", 1, 0, 5,
   "B", 1, 1, 4,
   "C", 1, 0, 6,
   "A", 2, 2, 3,
   "B", 2, 2, 4,
   "C", 2, 0, 6,
   "A", 3, 4, 2,
   "B", 3, 5, 3,
   "C", 3, 2, 7)
counts.to.reps(d, metamorphosed, did.not.metamorphose, competent)
}
