% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-plot.R
\name{example.plot}
\alias{example.plot}
\title{Create an example plot using the \code{\link[=gglogistic]{gglogistic()}} function.}
\usage{
example.plot()
}
\description{
Create an example plot using the \code{\link[=gglogistic]{gglogistic()}} function.
}
