% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogistic.R
\name{gglogistic}
\alias{gglogistic}
\title{Plot logistic regression of probability from 0\% to 100\% across a predictor.}
\usage{
gglogistic(
  data,
  predictor,
  success.counts,
  failure.counts,
  treatment,
  replicate,
  probability.of.interest = 0.5,
  xlabel = NULL,
  ylabel = NULL,
  treatment.label = NULL,
  treatment.colors = NULL,
  integer.x.breaks.by = NULL,
  geoms = "line",
  fill.alpha = 0.25,
  jitter.width = 0,
  dodge.width = 0
)
}
\arguments{
\item{data}{Data with continuous predictor and success and failure columns.}

\item{predictor}{Continuous predictor for x-axis.}

\item{success.counts}{Column representing number of successes per observation.}

\item{failure.counts}{Column representing number of failures per observation.}

\item{treatment}{Treatment variable.}

\item{replicate}{Within-treatment replicates. NOT IMPLEMENTED YET.}

\item{probability.of.interest}{Calculate predictor value in each group that
would yield this probability, or pass NULL to ignore.}

\item{xlabel}{Label for x-axis.}

\item{ylabel}{Label for y-axis.}

\item{treatment.label}{Label for treatment.}

\item{treatment.colors}{Vector mapping treatment levels to colors.}

\item{integer.x.breaks.by}{If specified, integer breaks in x-axis. Forces
ggplot to break at integer values.}

\item{geoms}{Vector containing any or all of c("line", "point", "boxplot").
"boxplot" only makes sense for data where the predictor is
conceptually continuous, but was only sampled at discrete steps.
For instance, if the predictor is time, and samples were only
taken once per day, a boxplot might be useful.}

\item{fill.alpha}{Opacity in range [0,1] for fill of boxplots. Boxplots can
be hard to see if there is a lot of data, and having a
light fill color helps to distinguish them.}

\item{jitter.width}{If specified, amount to jitter points in x-axis. Useful
if predictor was sampled at discrete steps that would
otherwise cause points to be hard to tell apart.}

\item{dodge.width}{If specified, amount to dodge replicate boxplots or groups
of points by.}
}
\value{
A ggplot object.
}
\description{
Plot logistic regression of probability from 0\% to 100\% across a predictor.
}
\examples{
logisticat::metamorphosis() \%>\%
  logisticat::gglogistic(
    predictor = days,
    success.counts = metamorphosed,
    failure.counts = did.not.metamorphose,
    treatment = ration,
    replicate = beaker,
    probability.of.interest = 0.5,
    xlabel = "Age (days)",
    ylabel = "Metamorphic Competence",
    treatment.label = "Algal Ration",
    treatment.colors = c("high" = "#f2af09","low" = "#ea1578"),
    geoms = c("line", "boxplot"))
}
