% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse-predict.R
\name{inverse.predict.by.replicate}
\alias{inverse.predict.by.replicate}
\title{Create a Data frame with an inverse prediction for each replicate.}
\usage{
inverse.predict.by.replicate(
  data,
  predictor,
  outcome,
  treatment,
  replicate,
  probability
)
}
\arguments{
\item{data}{A Data frame with a single row per individual}

\item{predictor}{Data variable to predict values in regressions.}

\item{outcome}{Data variable to represent outcomes.}

\item{treatment}{Data variable representing treatments.}

\item{replicate}{Data variable representing replicates within treatments.}

\item{probability}{Probability for which to make inverse predictions.}
}
\value{
A tibble with columns for replicate, treatment, and predictor
}
\description{
Create a Data frame with an inverse prediction for each replicate.
}
