% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse-predict.R
\name{inverse.predict}
\alias{inverse.predict}
\title{Find the value of a predictor variable that has the given probability.}
\usage{
inverse.predict(regression, probability)
}
\arguments{
\item{regression}{A logistic regression of the form \code{probability ~ predictor}
in a \code{glm} class object.}

\item{probability}{Value in range (0,1) for which predictor should be found.}
}
\value{
The value of the predictor variable for which the given probability
is predicted.
}
\description{
Find the value of a predictor variable that has the given probability.
}
