% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list.df.rows}
\alias{list.df.rows}
\title{Get a list a row vectors from a data frame.}
\usage{
list.df.rows(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
A list, with one vector for each row.
}
\description{
Get a list a row vectors from a data frame.
}
