% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-regression.R
\name{logistic.regression}
\alias{logistic.regression}
\title{Make a logistic regression for a binary outcome based on a continuous predictor.}
\usage{
logistic.regression(data, predictor, outcome)
}
\arguments{
\item{data}{\code{data.frame} with the columns named in the other arguments.}

\item{predictor}{Continuous predictor variable.}

\item{outcome}{\code{logical} outcome variable with values TRUE or FALSE.}
}
\value{
A logistic regression model in a \code{glm} class object.
}
\description{
Make a logistic regression for a binary outcome based on a continuous predictor.
}
