% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-geoms.R
\name{make.boxplot}
\alias{make.boxplot}
\title{Create a geom for a box plot of success fractions.}
\usage{
make.boxplot(fraction.data, x, y, group, fill.alpha = 0.25, ...)
}
\arguments{
\item{fraction.data}{Data frame with colum representing fraction of success.}

\item{x}{x-axis data variable, should be a predictor.}

\item{y}{y-axis data variable, fraction of successes in range [0,1].}

\item{group}{Data variable for groups and color.}

\item{fill.alpha}{Opacity of fill color for boxplot in range [0,1].}

\item{...}{Additional arguments for underlying geom.}
}
\value{
A geom of a boxplot.
}
\description{
Create a geom for a box plot of success fractions.
}
\keyword{internal}
