% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-geoms.R
\name{make.inverse.prediction.geoms}
\alias{make.inverse.prediction.geoms}
\title{Draw "T"-shaped intersection of predicted with probability.}
\usage{
make.inverse.prediction.geoms(
  inverse.prediction.df,
  predictor,
  probability,
  alpha = 1,
  digits = 3,
  ...
)
}
\arguments{
\item{inverse.prediction.df}{Data frame of invers predictions.}

\item{predictor}{Data variable for which inverse prediction was created.}

\item{probability}{Probabily for which inverse prediction was run.}

\item{alpha}{Opacity in range [0,1].}

\item{digits}{Digits of inverse predicted values to show.}

\item{...}{Additional arguments to geoms.}
}
\value{
A list of geoms.
}
\description{
Draw "T"-shaped intersection of predicted with probability.
}
