% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-geoms.R
\name{make.regression.line}
\alias{make.regression.line}
\title{Create a geom representing a logistic regression on the data.}
\usage{
make.regression.line(individual.rows, group, color, alpha = 1, ...)
}
\arguments{
\item{individual.rows}{Data frame with each row representing a pass or fail.}

\item{group}{Data variable specifying what groups to make into regressions.}

\item{color}{Data variable specifying how to color regression lines.}

\item{alpha}{Alpha value for line.}

\item{...}{Additional arguments to pass to the underlying geom.}
}
\value{
geom that can be added to a ggplot to show regressions.
}
\description{
Create a geom representing a logistic regression on the data.
}
\note{
x and y must be specified as aesthetics on the ggplot to which this is added.
}
\keyword{internal}
