% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{metamorphosis}
\alias{metamorphosis}
\title{Data for a hypothetical experiment on time to competence in echinoid larvae.}
\usage{
metamorphosis()
}
\value{
A \code{data.frame} with example data.
}
\description{
When exposed to potassium chloride, and echinoid larvae will either
metamorhpose, in which case it has reached metamorphic competence, or it will
fail to metamorphose, in which case it is not competent.
}
\details{
Note that the total of the "success" and "failure" counts don't need to
add up to the same value, they just happen to in this data set.
}
