% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nybble2hex}
\alias{nybble2hex}
\title{Get a hex string representation of a nybble}
\usage{
nybble2hex(nybble)
}
\arguments{
\item{nybble}{Integer value in range [0,15]}
}
\value{
String representation of nybble.
}
\description{
Get a hex string representation of a nybble
}
