% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome-counts-to-rows.R
\name{outcome.counts.to.rows}
\alias{outcome.counts.to.rows}
\title{Transform data with counts of binary outcomes into repeated rows.}
\usage{
outcome.counts.to.rows(data, success.counts, failure.counts, outcome)
}
\arguments{
\item{data}{\code{data.frame} with columns for success and failure counts.}

\item{success.counts}{Column with number of successes per group.}

\item{failure.counts}{Column with number of failures per group.}

\item{outcome}{Name to use for the outcome column.}
}
\value{
A \code{data.frame} with counts represented as repeated rows with a
number 1 representing a success or a 0 represnting a failure.
}
\description{
Given a \code{data.frame} containing results of a test with two possible outcomes
that is arranged with the number of "successes" per group in one column
and the number of "failures" per group in another column, return a new
\code{data.frame} that has a single outcome column with \code{1} for every success
and \code{0} for every failure, mapping each outcome to a single row.
}
\details{
The output is the format \code{glm()} expects when constructing logistic
regressions. The "successes" and "failures" can be any two mutually
exclusive states. Note that, while one would expect these to be logical
(TRUE and FALSE) values, \code{glm()} requires numeric values. You can think
about these numbers as representing the probability of success, which is
1 for success that actually happened and 0 for successes that did not.
}
