% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{sanity.check}
\alias{sanity.check}
\title{Small data set for use testing functions in this package.}
\usage{
sanity.check()
}
\value{
A \code{data.frame} with example data.
}
\description{
Small data set for use testing functions in this package.
}
\keyword{internal}
