% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with.alpha}
\alias{with.alpha}
\title{Set the alpha value of a color.}
\usage{
\method{with}{alpha}(color, alpha = 1)
}
\arguments{
\item{color}{Color specified as a name or hex string.}

\item{alpha}{Value in range [0,1] for alpha.}
}
\value{
A hex string representing the color with the chosen alpha.
}
\description{
Set the alpha value of a color.
}
\keyword{internal}
