% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogistic.R
\name{gglogistic}
\alias{gglogistic}
\title{Plot logistic regression of probability from 0\% to 100\% across a predictor.}
\usage{
gglogistic(
  data,
  predictor,
  success.counts,
  failure.counts,
  treatment,
  replicate = NULL,
  line.var = "both",
  point.var = NULL,
  boxplot.var = NULL,
  inverse.var = NULL,
  probability.of.interest = 0.5,
  xlabel = NULL,
  ylabel = NULL,
  treatment.label = NULL,
  treatment.colors = NULL,
  integer.x.breaks.by = NULL,
  fill.alpha = 0.25,
  replicate.alpha = 0.5,
  jitter.width = 0,
  dodge.width = 0
)
}
\arguments{
\item{data}{Data with continuous predictor and success and failure columns.}

\item{predictor}{Continuous predictor for x-axis.}

\item{success.counts}{Column representing number of successes per observation.}

\item{failure.counts}{Column representing number of failures per observation.}

\item{treatment}{Treatment variable.}

\item{replicate}{Within-treatment replicates.}

\item{line.var}{Variable to plot regression curves for. Must be the same as
\code{treatment} or \code{replicate}, or the character vector "both"
to plot both. Pass NULL to hide. Defaults to same variable
as \code{treatment}. y-axis represents predicted probability ad
each value of predictor.}

\item{point.var}{Variable to represent with points. Must be the same as
\code{treatment} or \code{replicate}, or the character vector "both"
to plot both. Pass NULL to hide. Defaults to NULL. y-axis
value represents fraction of successes in each observation
(for \code{replicate}) or across all observations with the
same treatment and same predictor (for \code{treatment}).}

\item{boxplot.var}{Variable to represent with points. Must be the same as
\code{treatment} or \code{replicate}, or the character vector "both"
to plot both. Pass NULL to hide. Defaults to NULL. y-axis
has same meaning as for \code{point.var}.}

\item{inverse.var}{Variable to draw lines representing inverse predictions
for. Must be the same as \code{treatment} or \code{replicate}, or
the character vector "both" to plot both. Pass NULL to
hide. Defaults to same variable as \code{line.var}.}

\item{probability.of.interest}{Calculate predictor value in each group that
would yield this probability.}

\item{xlabel}{Label for x-axis.}

\item{ylabel}{Label for y-axis.}

\item{treatment.label}{Label for treatment.}

\item{treatment.colors}{Vector mapping treatment levels to colors.}

\item{integer.x.breaks.by}{If specified, integer breaks in x-axis. Forces
ggplot to break at integer values.}

\item{fill.alpha}{Opacity in range [0,1] for fill of boxplots. Boxplots can
be hard to see if there is a lot of data, and having a
light fill color helps to distinguish them.}

\item{replicate.alpha}{Opacity in range [0,1] to draw \code{replicate}
geoms at when the same geom is also drawn for
\code{treatment}. Helps distinguish \code{treatment} from
\code{replicate} values.}

\item{jitter.width}{If specified, amount to jitter points in x-axis. Useful
if predictor was sampled at discrete steps that would
otherwise cause points to be hard to tell apart.}

\item{dodge.width}{If specified, amount to dodge replicate boxplots or groups
of points by.}
}
\value{
A ggplot object.
}
\description{
Plot logistic regression of probability from 0\% to 100\% across a predictor.
}
