% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome-counts-to-rows.R
\name{outcome.counts.to.rows}
\alias{outcome.counts.to.rows}
\title{Transform data with counts of binary outcomes into repeated rows.}
\usage{
outcome.counts.to.rows(data, success.counts, failure.counts, outcome)
}
\arguments{
\item{data}{\code{data.frame} with columns for success and failure counts.}

\item{success.counts}{Column with number of successes per group.}

\item{failure.counts}{Column with number of failures per group.}

\item{outcome}{Name to use for the outcome column.}
}
\value{
A \code{data.frame} with counts represented as repeated rows with a
\code{logical} flag representing a pass or fail.
}
\description{
Given a \code{data.frame} containing results of a test with two possible outcomes
that is arranged with the number of "successes" per group in one column
and the number of "failures" per group in another column, return a new
\code{data.frame} that has a single outcome column with TRUE and FALSE values,
with each outcome mapped to a single row.
}
\details{
The output is the format \code{glm()} expects when constructing logistic
regressions. The "successes" and "failures" can be any two mutually
exclusive states.
}
